+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - PR #5055 Add atan2 binary op
 - PR #5099 Add git commit hook for clang-format
 - PR #5072 Adding cython binding to `get_element`
+- PR #5092 Add `cudf::replace_nans`
 - PR #4881 Support row_number in rolling_window
 - PR #5068 Add Java bindings for arctan2
 
@@ -152,6 +153,7 @@
 - PR #5093 Add `.cat.as_known` related test in `dask_cudf`
 - PR #5106 Add detail API for `cudf::concatenate` with tables
 - PR #5104 Add missing `.inl` files to clang-format and git commit hook
+- PR #5112 Adding `htoi` and `ip2int` support to `StringMethods`
 - PR #5101 Add POSITION_INDEPENDENT_CODE flag to static cudftestutil library
 - PR #5109 Update CONTRIBUTING.md for `clang-format` pre-commit hook
 - PR #5054 Change String typecasting to be inline with Pandas
@@ -262,6 +264,7 @@
 - PR #5118 Fix naming for java string length operators
 - PR #5129 Fix missed reference in tests from 5118
 - PR #5122 Fix `clang-format` `custrings` bug
+- PR #5138 Install `contextvars` backport on Python 3.6
 
 
 # cuDF 0.13.0 (31 Mar 2020)